organization: move conversion logic to separate library
