+++ src/axom/inlet/tests/inlet_Reader.cpp
@@ -11,11 +11,7 @@
 
 #include "axom/config.hpp"
 
-#ifdef AXOM_USE_SOL
-  #include "axom/inlet/LuaReader.hpp"
-#endif
-
-#include "axom/inlet/YAMLReader.hpp"
+#include "axom/inlet/tests/inlet_test_utils.hpp"
 
 template <typename InletReader>
 class InletReaderTest : public testing::Test
@@ -23,143 +19,12 @@ class InletReaderTest : public testing::Test
 
 TYPED_TEST_SUITE_P(InletReaderTest);
 
-namespace detail
-{
-class LuaToYAML
-{
-  static std::vector<std::string> tokenize(const std::string& text)
-  {
-    std::vector<std::string> result;
-    std::size_t pos = 0;
-    auto add_to_result = [&result](std::string&& str) {
-      std::vector<std::string> parsed_after;
-      std::size_t pos;
-      while(str.length() > 1 &&
-            ((pos = str.find_first_of("{}[];,")) != std::string::npos))
-      {
-        if(pos == 0)
-        {
-          result.emplace_back(1, str[pos]);
-          str = str.substr(1, str.length() - 1);
-        }
-        else
-        {
-          parsed_after.emplace_back(1, str[pos]);
-          str = str.substr(0, str.length() - 1);
-        }
-      }
-      if(!str.empty())
-      {
-        result.push_back(str);
-      }
-      result.insert(result.end(), parsed_after.rbegin(), parsed_after.rend());
-    };
-
-    while(pos < text.length())
-    {
-      if(text[pos] == '"')
-      {
-        auto ending_quote = text.find('"', pos + 1);
-        add_to_result(text.substr(pos, ending_quote - pos + 1));
-        pos = ending_quote + 1;
-      }
-      else if(text[pos] == '[')
-      {
-        auto ending_bracket = text.find(']', pos + 1);
-        add_to_result(text.substr(pos, ending_bracket - pos + 1));
-        pos = ending_bracket + 1;
-      }
-      else
-      {
-        auto ending_space = text.find(' ', pos + 1);
-        if(ending_space == std::string::npos)
-        {
-          if(text[pos] == ' ')
-          {
-            pos++;
-          }
-          add_to_result(text.substr(pos, text.length() - pos));
-          pos = text.length();
-        }
-        else
-        {
-          add_to_result(text.substr(pos, ending_space - pos));
-          pos = ending_space + 1;
-        }
-      }
-    }
-    return result;
-  }
-
-public:
-  static std::string convert(const std::string& luaString)
-  {
-    const auto tokens = tokenize(luaString);
-    std::size_t i = 0;
-    std::string indent = "";
-    std::string result;
-    while(i < tokens.size())
-    {
-      const auto& token = tokens[i];
-      if((i < tokens.size() - 2) && tokens[i + 1] == "=" && tokens[i + 2] == "{")
-      {
-        result += indent + token + ":\n";
-        indent += "  ";
-        i += 3;
-      }
-      else if(token == "}")
-      {
-        indent = indent.substr(2);
-        i += 1;
-      }
-      else if(token == "," || token == ";")
-      {
-        i += 1;
-      }
-      else if(token == "[")
-      {
-        const auto& key = tokens[i + 1];
-
-        const auto& value = tokens[i + 4];
-        if(isdigit(key.front()))
-        {
-          // Assume this is an integer-keyed array
-          result += indent + "- " + value + "\n";
-        }
-        else
-        {
-          result += indent + key + ": " + value + "\n";
-        }
-        i += 5;
-      }
-      else
-      {
-        const auto& value = tokens[i + 2];
-        result += indent + token + ": " + value + "\n";
-        i += 3;
-      }
-    }
-    return result;
-  }
-};
-
-template <typename InletReader>
-std::string fromLuaTo(const std::string& luaString)
-{
-  return luaString;
-}
-
-template <>
-std::string fromLuaTo<axom::inlet::YAMLReader>(const std::string& luaString)
-{
-  return LuaToYAML::convert(luaString);
-}
-}  // namespace detail
+using axom::inlet::detail::fromLuaTo;
 
 TYPED_TEST_P(InletReaderTest, getTopLevelBools)
 {
   TypeParam reader;
-  reader.parseString(detail::fromLuaTo<TypeParam>("foo = true; bar = false"));
+  reader.parseString(fromLuaTo<TypeParam>("foo = true; bar = false"));
 
   bool value, retValue;
 
@@ -177,8 +42,7 @@ TYPED_TEST_P(InletReaderTest, getTopLevelBools)
 TYPED_TEST_P(InletReaderTest, getInsideBools)
 {
   TypeParam reader;
-  reader.parseString(
-    detail::fromLuaTo<TypeParam>("foo = { bar = false; baz = true }"));
+  reader.parseString(fromLuaTo<TypeParam>("foo = { bar = false; baz = true }"));
 
   bool value, retValue;
 
@@ -196,7 +60,7 @@ TYPED_TEST_P(InletReaderTest, getInsideBools)
 TYPED_TEST_P(InletReaderTest, getTopLevelStrings)
 {
   TypeParam reader;
-  reader.parseString(detail::fromLuaTo<TypeParam>(
+  reader.parseString(fromLuaTo<TypeParam>(
     "foo = \"this is a test string\"; bar = \"TesT StrInG\""));
 
   bool retValue;
@@ -216,7 +80,7 @@ TYPED_TEST_P(InletReaderTest, getTopLevelStrings)
 TYPED_TEST_P(InletReaderTest, getInsideStrings)
 {
   TypeParam reader;
-  reader.parseString(detail::fromLuaTo<TypeParam>(
+  reader.parseString(fromLuaTo<TypeParam>(
     "foo = { bar = \"this is a test string\"; baz = \"TesT StrInG\" }"));
 
   bool retValue;
@@ -236,7 +100,7 @@ TYPED_TEST_P(InletReaderTest, getInsideStrings)
 TYPED_TEST_P(InletReaderTest, mixLevelTables)
 {
   TypeParam reader;
-  reader.parseString(detail::fromLuaTo<TypeParam>(
+  reader.parseString(fromLuaTo<TypeParam>(
     "t = { innerT = { foo = 1 }, anotherInnerT = {baz = 3}}"));
 
   bool retValue;
@@ -265,7 +129,7 @@ TYPED_TEST_P(InletReaderTest, getMap)
     "luaArray = { [0] = 4, [1] = 5, [2] = 6 , [3] = true, [4] = false, [5] = "
     "2.4, [6] = 'hello', [7] = 'bye' }";
   TypeParam reader;
-  reader.parseString(detail::fromLuaTo<TypeParam>(testString));
+  reader.parseString(fromLuaTo<TypeParam>(testString));
 
   std::unordered_map<int, int> ints;
   bool found = reader.getIntMap("luaArray", ints);
